Commit da alteração do ex007d2 e da realização do ex007d6

+++ ex007d2.py
@@ -1,4 +1,4 @@
-n1 = int(input('Digite o numero em questão: '))
+n1 = float(input('Digite o numero em questão: '))
 dobro = n1 * 2
 triplo = n1 * 3
 raizQuad = n1 **(1/2)
